Move argument out of group.
The manager name argument should not be
inside the group itself. Having it in this
location causes an error where it cannot be
assigned via launch file (command line
assignment seems to work fine, so presumably
this is a bug in roslaunch). Moving the argument
out of the group appears to work fine in both
launch files and via command line.